+++ src/main/java/org/example/arraytask/factory/ArrayFactory.java
@@ -2,6 +2,8 @@ package org.example.arraytask.factory;
 
 import org.example.arraytask.entity.ArrayWrapper;
 
+import java.math.BigInteger;
+
 public interface ArrayFactory {
-    ArrayWrapper create(int[] values, String name);
+    ArrayWrapper create(BigInteger[] values, String name);
 }
