+++ src/main/java/org/example/arraytask/factory/IntegerArrayFactory.java
@@ -2,9 +2,11 @@ package org.example.arraytask.factory;
 
 import org.example.arraytask.entity.ArrayWrapper;
 
+import java.math.BigInteger;
+
 public class IntegerArrayFactory implements ArrayFactory {
     @Override
-    public ArrayWrapper create(int[] values, String name) {
+    public ArrayWrapper create(BigInteger[] values, String name) {
         return new ArrayWrapper(values, name);
     }
 }
